requirements.txt
pynput
psutil
google-cloud-translate
grpcio-tools
python-lsp-server[all]
langchain
